+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+New in 1.18.0
+=============
+
+* Add configuration parameter `low_mapq` to set what you consider as low
+  mapping quality. Default is 5 which is the setting hardcoded until now.
+
 New in 1.17.0
 =============